rename a bunch of the files
