+++ create_manifest/handler.py
